+++ src/main/java/com/sqli/testapp/service/imp/OrderServiceImp.java
@@ -7,6 +7,7 @@ import com.sqli.testapp.model.Order;
 import com.sqli.testapp.repository.ClientRepository;
 import com.sqli.testapp.repository.OrderRepository;
 import com.sqli.testapp.service.OrderService;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -34,7 +35,7 @@ public class OrderServiceImp implements OrderService {
     public Order storeOrder(OrderDto orderDto) {
         Optional<Client> clientOptional = clientRepository.findById(orderDto.getClientId());
         if (!clientOptional.isPresent()) {
-            throw new IllegalArgumentException("Client not found with id: " + orderDto.getClientId());
+            throw new EntityNotFoundException("Client not found with id: " + orderDto.getClientId());
         }
         Order order = orderMapper.dtoRequestToEntity(orderDto);
         order.setClient(clientOptional.get());
@@ -46,11 +47,11 @@ public class OrderServiceImp implements OrderService {
     public Order updateOrder(int id, OrderDto orderDto) {
         Optional<Order> orderOptional = orderRepository.findById(id);
         if(!orderOptional.isPresent()){
-            throw new IllegalArgumentException("Order not found with id: " + id);
+            throw new EntityNotFoundException("Order not found with id: " + id);
         }
         Optional<Client> clientOptional = clientRepository.findById(orderDto.getClientId());
         if (!clientOptional.isPresent()) {
-            throw new IllegalArgumentException("Client not found with id: " + orderDto.getClientId());
+            throw new EntityNotFoundException("Client not found with id: " + orderDto.getClientId());
         }
         Order order = orderOptional.get();
         order.setAmount(orderDto.getAmount());
@@ -60,8 +61,14 @@ public class OrderServiceImp implements OrderService {
     }
 
     @Override
-    public void removeOrder(int id) {
+    public boolean removeOrder(int id) {
+        Optional<Order> orderOptional = orderRepository.findById(id);
+        if (!orderOptional.isPresent()) {
+            return false;
+        }
         orderRepository.deleteById(id);
+        return true;
     }
+
 
 }
